Bring back squisher
Some devices could benefit from the size reduction from running
squisher, so bring back a revised version as an optional part of the
build process.

This version of squisher no longer handles the process of making
release OTAs, instead operating on $OUT/system before system images and
OTAs are created.  This has the advantage of avoiding having to unzip a
just-packed OTA to optimize it before repacking it.

Change-Id: I288f694fd1bf26782e05dcfa797a0569d9fdf9cb